Fix my utter stupidity. Poop
